requirements.txt
# Core dependencies
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
pyarrow>=14.0.0  # For parquet support

# LLM and NLP
openai>=1.0.0
anthropic>=0.18.0

# Visualization
matplotlib>=3.7.0
plotly>=5.17.0
seaborn>=0.12.0

# WebSocket and web server
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# Audio processing (for voice input)
# Note: pyaudio may require system dependencies (portaudio)
# On macOS: brew install portaudio
# On Ubuntu: sudo apt-get install portaudio19-dev
# pyaudio>=0.2.13  # Uncomment if you need real audio processing
speechrecognition>=3.10.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
numpy>=1.24.0

# Code execution safety
# Note: ast is a built-in Python module, no need to install
executing>=2.0.0

